Move IProxyConfigurationListener
