+++ src/SLCore/Listener/Proxy/IProxyConfigurationListener.cs
@@ -18,19 +18,13 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.ComponentModel.Composition;
 using System.Threading.Tasks;
 using SonarLint.VisualStudio.SLCore.Core;
 using SonarLint.VisualStudio.SLCore.Listener.Proxy;
 
 namespace SonarLint.VisualStudio.SLCore.Listeners.Implementation;
 
-[Export(typeof(ISLCoreListener))]
-[PartCreationPolicy(CreationPolicy.Shared)]
-public class ProxyConfigurationListener : IProxyConfigurationListener
+public interface IProxyConfigurationListener : ISLCoreListener
 {
-    public Task<SelectProxiesResponse> SelectProxiesAsync(object parameters)
-    {
-        return Task.FromResult(new SelectProxiesResponse());
-    }
+    Task<SelectProxiesResponse> SelectProxiesAsync(object parameters);
 }
